% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-data.R
\name{occ_daily_volume}
\alias{occ_daily_volume}
\title{Daily volumes by date and API}
\usage{
occ_daily_volume(api.filter = NULL, exclude = FALSE, after = NULL,
  ...)
}
\arguments{
\item{api.filter}{character; API numbers to filter}

\item{exclude}{logical; should \code{api.filter} by exclusionary? The default (\code{FALSE})
will return only API numbers specified in \code{api.filter}.}

\item{after}{date or object coercible to date; A non-null entry with return results after
the specified; otherwise all data are returned.}

\item{...}{additional paramters sent to \code{\link{dcast}}}
}
\description{
Daily volumes by date and API
}
\examples{
\dontrun{
#
# Plot total injection across Oklahoma
# (complete since early 2015)
#

require(dplyr)

dv <- occ_daily_volume()
# each API gets a column in the result
# and date is given in `Report.Date`
head(names(dv))

Dates <- dv$Report.Date
Vols <- dplyr::select(dv,-Report.Date)

plot(Dates, rowSums(Vols, na.rm=TRUE), type='l')

}
}
